requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0

# Medical imaging libraries
TIAtoolbox>=1.7.0
openslide-python>=1.2.0
histomicstk>=1.3.0

# Image processing
opencv-python>=4.8.0
Pillow>=9.5.0
scikit-image>=0.20.0

# Scientific computing
scikit-learn>=1.3.0
scipy>=1.11.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Experiment tracking
tensorboard>=2.13.0
wandb>=0.15.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.65.0
click>=8.1.0

# Quality and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Medical imaging specific
pydicom>=2.4.0
SimpleITK>=2.3.0